+++ tortoise/migrations/expressions.py
@@ -5,5 +5,5 @@ class RawSQLTerm:
     def __init__(self, sql: str) -> None:
         self.sql = sql
 
-    def get_sql(self, _context) -> str:
+    def get_sql(self, _context=None, dialect: str | None = None) -> str:
         return self.sql